refactor: 함수로 추출, 불필요한 헤더 제거, 검증 로직 추가

+++ src/main/java/com/elice/bookstore/user/service/UserService.java
@@ -1,21 +1,30 @@
 package com.elice.bookstore.user.service;
 
-import com.elice.bookstore.config.exception.domain.user.*;
+import com.elice.bookstore.config.exception.domain.user.UserDuplicatedEmailException;
+import com.elice.bookstore.config.exception.domain.user.UserEmptyPasswordException;
+import com.elice.bookstore.config.exception.domain.user.UserNotAuthorizedException;
+import com.elice.bookstore.config.exception.domain.user.UserNotExistException;
+import com.elice.bookstore.config.exception.domain.user.UserNotMatchPasswordException;
+import com.elice.bookstore.config.exception.domain.user.UserShortPasswordException;
 import com.elice.bookstore.config.security.authentication.user.CustomUserDetails;
 import com.elice.bookstore.user.domain.Role;
 import com.elice.bookstore.user.domain.User;
-import com.elice.bookstore.user.dto.*;
+import com.elice.bookstore.user.dto.CustomSecurityContext;
+import com.elice.bookstore.user.dto.RequestDeleteUser;
+import com.elice.bookstore.user.dto.RequestModifyUser;
+import com.elice.bookstore.user.dto.RequestRegisterUser;
+import com.elice.bookstore.user.dto.ResponseLookupUser;
+import com.elice.bookstore.user.dto.ResponseRegisterUser;
 import com.elice.bookstore.user.mapper.UserMapper;
 import com.elice.bookstore.user.repository.UserRepository;
+import java.util.Collection;
+import java.util.Iterator;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Collection;
-import java.util.Iterator;
 
 /**
  * userService [signup].
@@ -80,28 +89,23 @@ public class UserService {
     return userMapper.UserToResponseRegisterUser(savedUser);
   }
 
+  /**
+   * lookup user.
+   */
   public ResponseLookupUser lookup(String separator) {
 
-    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-    CustomUserDetails customUserDetails = (CustomUserDetails) authentication.getPrincipal();
-
-    Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-    Iterator<? extends GrantedAuthority> iterator = authorities.iterator();
-    String auth = iterator.next().getAuthority();
+    CustomSecurityContext sc = getCustomSecurityContext();
 
     User user = null;
-    if (auth.equals("USER")) {
+    if (sc.auth().equals("USER")) {
       if (!separator.equals("me")) {
         throw new UserNotAuthorizedException();
       }
-      long id = Long.parseLong(customUserDetails.getId());
-      user = userRepository.findByIdAndIsExist(id, true).orElseThrow(
+      user = userRepository.findByIdAndIsExist(sc.id(), true).orElseThrow(
           UserNotExistException::new
       );
-    } else if (auth.equals("ADMIN")) {
-      Long id = Long.parseLong(separator);
-      user = userRepository.findByIdAndIsExist(id, true).orElseThrow(
+    } else if (sc.auth().equals("ADMIN")) {
+      user = userRepository.findByIdAndIsExist(sc.id(), true).orElseThrow(
           UserNotExistException::new
       );
     }
@@ -109,9 +113,7 @@ public class UserService {
     return userMapper.UserToResponseLookupUser(user);
   }
 
-  @Transactional
-  public ResponseLookupUser modify(String separator, RequestModifyUser requestModifyUser) {
-
+  private CustomSecurityContext getCustomSecurityContext() {
     Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
     CustomUserDetails customUserDetails = (CustomUserDetails) authentication.getPrincipal();
@@ -120,52 +122,71 @@ public class UserService {
     Iterator<? extends GrantedAuthority> iterator = authorities.iterator();
     String auth = iterator.next().getAuthority();
 
+    long id = Long.parseLong(customUserDetails.getId());
+
+    return new CustomSecurityContext(id, auth);
+  }
+
+  /**
+   * modify user.
+
+   * @param separator two cases {me} - Role:USER, {1, 2, 3, 4, ...} - Role:ADMIN.
+   * @param requestModifyUser .
+   * @return user info.
+   */
+  @Transactional
+  public ResponseLookupUser modify(String separator, RequestModifyUser requestModifyUser) {
+
+    CustomSecurityContext sc = getCustomSecurityContext();
+
     User user = null;
-    if (auth.equals("USER")) {
+    if (sc.auth().equals("USER")) {
       if (!separator.equals("me")) {
         throw new UserNotAuthorizedException();
       }
-      long id = Long.parseLong(customUserDetails.getId());
-      user = userRepository.findByIdAndIsExist(id, true).orElseThrow(
+      user = userRepository.findByIdAndIsExist(sc.id(), true).orElseThrow(
           UserNotExistException::new
       );
-    } else if (auth.equals("ADMIN")) {
-      Long id = Long.parseLong(separator);
-      user = userRepository.findByIdAndIsExist(id, true).orElseThrow(
+    } else if (sc.auth().equals("ADMIN")) {
+      user = userRepository.findByIdAndIsExist(sc.id(), true).orElseThrow(
           UserNotExistException::new
       );
     }
 
+    assert user != null;
     user.modifyUser(requestModifyUser);
 
     return userMapper.UserToResponseLookupUser(user);
   }
 
+  /**
+   * delete user.
+
+   * @param separator .
+   * @param requestDeleteUser .
+   */
   @Transactional
   public void delete(String separator, RequestDeleteUser requestDeleteUser) {
 
-    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-    CustomUserDetails customUserDetails = (CustomUserDetails) authentication.getPrincipal();
-
-    Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-    Iterator<? extends GrantedAuthority> iterator = authorities.iterator();
-    String auth = iterator.next().getAuthority();
+    CustomSecurityContext sc = getCustomSecurityContext();
 
     User user = null;
-    if (auth.equals("USER")) {
+    if (sc.auth().equals("USER")) {
       if (!separator.equals("me")) {
         throw new UserNotAuthorizedException();
       }
-      long id = Long.parseLong(customUserDetails.getId());
-      user = userRepository.findByIdAndIsExist(id, true).orElseThrow(
+      user = userRepository.findByIdAndIsExist(sc.id(), true).orElseThrow(
           UserNotExistException::new
       );
-    } else if (auth.equals("ADMIN")) {
-      Long id = Long.parseLong(separator);
-      user = userRepository.findByIdAndIsExist(id, true).orElseThrow(
+    } else if (sc.auth().equals("ADMIN")) {
+      user = userRepository.findByIdAndIsExist(sc.id(), true).orElseThrow(
           UserNotExistException::new
       );
+    }
+
+    assert user != null;
+    if (!bcryptPasswordEncoder.matches(requestDeleteUser.password(), user.getPassword())) {
+      throw new UserNotMatchPasswordException();
     }
 
     user.deleteUser();
